fix uint

+++ board/safety/safety_hyundai_canfd.h
@@ -93,10 +93,10 @@ static addr_checks build_canfd_addr_checks(void) {
     return (addr_checks){base_addr_checks, BASE_ADDR_CHECK_LEN};
   }
 
-  for (int i = 0; i < BASE_ADDR_CHECK_LEN; i++) {
+  for (uint8_t i = 0; i < BASE_ADDR_CHECK_LEN; i++) {
     new_addresses[i] = base_addr_checks[i];
   }
-  
+
   if (hyundai_camera_scc) {
     new_addresses[BASE_ADDR_CHECK_LEN] = (AddrCheckStruct){.msg = {{0x1a0, 0, 32, .check_checksum = true, .max_counter = 0xffU, .expected_timestep = 20000U}, { 0 }, { 0 }}};
   } else {
